+++ web/.deprecated/app/app/uptime/blog/page.tsx
@@ -1,6 +1,0 @@
-import BlogPage from "@/components/uptime/allPages/BlogPage"
-
-
-const Page = () => <BlogPage />
-  
-export default Page