+++ web/src/app/dashboard/uptime/api/page.tsx
@@ -1,0 +1,6 @@
+import ApiPage from "@/components/uptime/all-pages/ApiPage"
+
+
+const Page = () => <ApiPage />
+  
+export default Page