+++ web/src/app/dashboard/uptime/calendar/page.tsx
@@ -1,0 +1,5 @@
+import CalendarMainPage from "@/components/uptime/calendar/main-page/MainPage";
+
+const Page = () => <CalendarMainPage />;
+
+export default Page;