+++ web/src/app/dashboard/uptime/layout.tsx
@@ -1,0 +1,17 @@
+import Header from "@/components/shared-layouts/header/Header"
+
+
+const UptimeLayout = ({
+    children,
+  }: {
+    children: React.ReactNode;
+  }) => {
+    return (
+      <>
+        <Header />
+        {children}
+      </>
+    )
+  }
+  
+export default UptimeLayout