+++ web/src/app/dashboard/uptime/server/page.tsx
@@ -1,0 +1,6 @@
+import ServerPage from "@/components/uptime/all-pages/ServerPage"
+
+
+const Page = () => <ServerPage />
+  
+export default Page